requirements.txt
fitz==1.19.6
pytesseract==0.3.8
Pillow==8.4.0
openpyxl==3.0.9
ocrmypdf==12.0.0